Moved a few pipes out of shared module into where they're used
