+++ src/app/news/news.module.ts
@@ -8,6 +8,8 @@ import { LatestNewsComponent } from './latest-news/latest-news.component';
 import { SharedModule } from '../shared/shared.module';
 import { FeaturedNewsComponent } from './featured-news/featured-news.component';
 import { NewsRoutingModule } from './news-routing.module';
+import { IsoDatePipe } from './pipes/iso-date.pipe';
+import { FormattedDatePipe } from './pipes/formatted-date.pipe';
 
 @NgModule({
   imports: [
@@ -16,7 +18,14 @@ import { NewsRoutingModule } from './news-routing.module';
     NewsRoutingModule,
     SharedModule
   ],
-  declarations: [NewsComponent, NewsArticleComponent, LatestNewsComponent, FeaturedNewsComponent],
+  declarations: [
+    NewsComponent,
+    NewsArticleComponent,
+    LatestNewsComponent,
+    FeaturedNewsComponent,
+    IsoDatePipe,
+    FormattedDatePipe
+  ],
   exports: [
     LatestNewsComponent
   ]
