+++ src/app/profile/profiles.component.ts
@@ -4,7 +4,7 @@ import { ActivatedRoute } from '@angular/router';
 import { Host } from './host';
 import { BreadcrumbConfigItem } from '../shared/breadcrumb/breadcrumb-config-item';
 import { profilesConfigActive } from '../shared/breadcrumb/breadcrumb-config';
-import { SortOrder } from '../shared/pipes/sort-order';
+import { SortOrder } from './pipes/sort-order';
 import { BreadcrumbService } from '../shared/services/breadcrumb/breadcrumb.service';
 
 @Component({
