--- src/app/shared/pipes/map-to-array.pipe.ts
@@ -1,18 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-import { DomSanitizer } from '@angular/platform-browser';
-
-// https://forum.ionicframework.com/t/inserting-html-via-angular-2-use-of-domsanitizationservice-bypasssecuritytrusthtml/62562/2
-@Pipe({
-    name: 'mapToArray'
-})
-export class MapToArrayPipe implements PipeTransform {
-  constructor(protected _sanitizer: DomSanitizer) {}
-
-  public transform(value, args: string[]): any {
-    const arr = [];
-    for (const key of Object.keys(value)) {
-      arr.push({key: key, value: value[key]});
-    }
-    return arr;
-  }
-}
