+++ src/app/shared/shared.module.ts
@@ -38,11 +38,7 @@ import { HttpRequestService } from './services/http-request/http-request.service
 
 // Pipes
 import { SafePipe } from './pipes/safe.pipe';
-import { SortByPipe } from './pipes/sort-by.pipe';
-import { MapToArrayPipe } from './pipes/map-to-array.pipe';
 import { TimePipe } from './pipes/time.pipe';
-import { FormattedDatePipe } from './pipes/formatted-date.pipe';
-import { IsoDatePipe } from './pipes/iso-date.pipe';
 
 // Directives
 import { ActiveDirective } from './directives/active.directive';
@@ -75,11 +71,7 @@ export function customHttpLoader(http: HttpClient) {
     DonateComponent,
     SocialLinksComponent,
     SafePipe,
-    SortByPipe,
-    MapToArrayPipe,
     TimePipe,
-    FormattedDatePipe,
-    IsoDatePipe,
     ContentBoxComponent,
     RadioPlayerComponent,
     LoadingSpinnerComponent,
@@ -105,11 +97,7 @@ export function customHttpLoader(http: HttpClient) {
     DonateComponent,
     SocialLinksComponent,
     SafePipe,
-    SortByPipe,
-    MapToArrayPipe,
     TimePipe,
-    FormattedDatePipe,
-    IsoDatePipe,
     ContentBoxComponent,
     RadioPlayerComponent,
     LoadingSpinnerComponent,
